roachtest: move load groups to util file
That way they can move with the tests into `tests/` when the moment
comes.

Release note: None
